requirements.txt
python-telegram-bot==20.7
opencv-python==4.9.0.80
pydub==0.25.1
